# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/countries": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Create country",
                "parameters": [
                    {
                        "description": "country",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.CreateUpdateCountryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "create a country",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.CountryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                        }
                    },
                    "409": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/countries/get-by-filter": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "GetByFilter country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "GetByFilter country",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get country by id",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.PagedList-github_com_alireza-fa_golang-car-shop_api_dto_CountryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/countries/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "GetById country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "GetById country",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "country id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get country by id",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.CountryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "204": {
                        "description": "get a country"
                    },
                    "409": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Delete country",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "country id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "delete a country",
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                        }
                    },
                    "409": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Update country",
                "parameters": [
                    {
                        "description": "country request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.CreateUpdateCountryRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "country id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "update a country",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.CountryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                        }
                    },
                    "409": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_alireza-fa_golang-car-shop_api_dto.CityResponse": {
            "type": "object",
            "properties": {
                "country": {
                    "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.CountryResponse"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_alireza-fa_golang-car-shop_api_dto.CountryResponse": {
            "type": "object",
            "properties": {
                "cities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.CityResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_alireza-fa_golang-car-shop_api_dto.CreateUpdateCountryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                }
            }
        },
...
        "github_com_alireza-fa_golang-car-shop_api_dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "description": "text number",
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "description": "contains notContains equals notEqual startWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endWith",
                    "type": "string"
                }
            }
        },
...
        "github_com_alireza-fa_golang-car-shop_api_dto.PagedList-github_com_alireza-fa_golang-car-shop_api_dto_CountryResponse": {
            "type": "object",
            "properties": {
                "hasNextPage": {
                    "type": "boolean"
                },
                "hasPreviousPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.CountryResponse"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alireza-fa_golang-car-shop_api_dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_dto.Sort"
                    }
                }
            }
        },
...
        "github_com_alireza-fa_golang-car-shop_api_dto.Sort": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        },
...
        "github_com_alireza-fa_golang-car-shop_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alireza-fa_golang-car-shop_api_validations.ValidationError"
                    }
                }
            }
        },
        "github_com_alireza-fa_golang-car-shop_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                442902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "github_com_alireza-fa_golang-car-shop_api_validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}